separate logic for  normal and emulator publisher
